pyproject.toml
[project]
name = "write_a_book_with_flows"
version = "0.1.0"
description = "write_a_book_with_flows using crewAI"
authors = [
    { name = "Your Name", email = "you@example.com" },
]
requires-python = ">=3.10,<=3.13"
dependencies = [
    "crewai[tools]>=0.152.0",
    "python-dotenv",
    "opentelemetry-instrumentation-openai",
    "opentelemetry-instrumentation-crewai",
    "braintrust[otel]",
]

[project.scripts]
kickoff = "write_a_book_with_flows.main:kickoff"
plot = "write_a_book_with_flows.main:plot"

[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
